% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{E_y}
\alias{E_y}
\title{E_y}
\usage{
E_y(X, W, cure_res, t_new, J)
}
\arguments{
\item{X}{A vector of covariates for the longitudinal outcome.}

\item{W}{A vector of covariates for the survival outcome.}

\item{cure_res}{A list containing the estimated parameters of the cure rate model.}

\item{t_new}{A new time point for which the expected value is calculated.}

\item{J}{The number of Monte Carlo samples used for the calculation.}
}
\value{
The expected value of the longitudinal outcome given the cure rate model parameters.
}
\description{
This function calculates the expected value of the longitudinal outcome given the cure rate model parameters for a new time point.
}
\examples{
 # Make sure that you have run the examples from infer_CRJoint_MLE
 X_bar <- mean(survdat$Y0SCALE)
 W_bar <- mean(survdat$Y0SCALE)
 E_y(X_bar, W_bar, results, 0.1, 100)
}
