% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{infer_CRJoint_MLE}
\alias{infer_CRJoint_MLE}
\title{infer_CRJoint_MLE}
\usage{
infer_CRJoint_MLE(
  survdat,
  longdat,
  init,
  fmla.tte,
  fmla.long,
  longdat.time,
  id.indicator,
  maxIter = 5000,
  tol = 8e-04,
  no_cure = FALSE
)
}
\arguments{
\item{survdat}{A data frame containing the survival data.}

\item{longdat}{A data frame containing the longitudinal data.}

\item{init}{A list containing the initial values for the parameters.}

\item{fmla.tte}{A formula specifying the survival outcome.}

\item{fmla.long}{A formula specifying the longitudinal outcome.}

\item{longdat.time}{A string specifying the time variable in the longitudinal data.}

\item{id.indicator}{A string specifying the ID variable in the data.}

\item{maxIter}{An integer specifying the maximum number of iterations for the EM algorithm.}

\item{tol}{A numeric value specifying the convergence tolerance for the EM algorithm.}

\item{no_cure}{A logical value indicating whether to include the cure rate model.}
}
\value{
A list containing the estimated parameters of the joint cure rate model.
}
\description{
This function estimates the parameters of the joint cure rate model using the EM algorithm.
}
\examples{
  # Load required library
  library(JoCuR)

  # Set seed for reproducibility
  set.seed(123)

  # Simulate example data
  sim_data <- CRsimulation(
    fmla.tte = as.formula(Surv(PFS_YEARS, PFS_EVENT) ~ 0 + Y0SCALE),
    fmla.long = as.formula(PCHG ~ 0 + Y0SCALE),
    beta.tte = c(0.2), normal.tte = TRUE, sd.tte = 0.2,
    beta.y = c(-0.5), sd.y = 0.1, beta.y.cure = c(-0.2), sd.y.cured = 0.2,
    cured.rate = 0.4, cured.mean = c(0, -0.2), cured.sd = c(0.2, 0.2),
    cured.corr = matrix(c(1, 0.5, 0.5, 1), nrow = 2),
    randeff.mean = c(0.5, 0, -0.5, 0.5), randeff.sd = rep(0.2, 4),
    randeff.corr = matrix(c(1, -0.4, -0.2, -0.3,
                           -0.4, 1, 0.5, 0.20,
                           -0.2, 0.5, 1, 0.2,
                           -0.3, 0.20, 0.2, 1), nrow = 4),
    n = 100, censor.parameter = 0.5, time.interval = 0.1
  )

  # Extract datasets
  longdat <- sim_data[["longdat"]]
  survdat <- sim_data[["survdat"]]

  # Define formulas
  fmla.tte <- as.formula(Surv(PFS_YEARS, PFS_EVENT) ~ 0 + Y0SCALE)
  fmla.long <- as.formula(PCHG ~ 0 + Y0SCALE)

  # Initial values
  init <- list(
    mu_r = c(0.5, 0, -0.5, 0.5),
    Sigma_r = diag(rep(0.2, 4)) \%*\% 
              matrix(c(1, -0.4, -0.2, -0.3,
                      -0.4, 1, 0.5, 0.20,
                      -0.2, 0.5, 1, 0.2,
                      -0.3, 0.20, 0.2, 1), nrow = 4) \%*\% 
              diag(rep(0.2, 4)),
    beta_y = c(-0.5),
    sigma_y_sq = 0.01,
    beta_tte = c(0.2),
    sigma_tte_sq = 0.04,
    pi_c = 0.4,
    beta_cure = c(-0.2),
    mu_r_cure = c(0, -0.2),
    Sigma_r_cure = diag(c(0.2, 0.2)) \%*\% 
                   matrix(c(1, 0.5, 0.5, 1), nrow = 2) \%*\% 
                   diag(c(0.2, 0.2)),
    sigma_y_cure_sq = 0.04
  )

  # Fit the joint model
  results <- infer_CRJoint_MLE(
    survdat = survdat,
    longdat = longdat,
    init = init,
    fmla.tte = fmla.tte,
    fmla.long = fmla.long,
    longdat.time = "visittime",
    id.indicator = "id",
    maxIter = 50,
    tol = 5e-3,
    no_cure = FALSE
  )
}
